restructure actions
